+++ database/migrations/2020_01_01_000007_create_attribute_entity_table.php
@@ -17,9 +17,9 @@ class CreateAttributeEntityTable extends Migration
     {
         Schema::create(config('rinvex.attributes.tables.attribute_entity'), function (Blueprint $table) {
             // Columns
-            $table->integer('attribute_id')->unsigned();
+            $table->bigInteger('attribute_id')->unsigned();
             $table->string('entity_type');
-            $table->integer('entity_id')->unsigned()->nullable(); // TODO: Making this nullable for now as it breaks the basic features
+            $table->bigInteger('entity_id')->unsigned()->nullable(); // TODO: Making this nullable for now as it breaks the basic features
             $table->timestamps();
 
             // Indexes
